+++ QoolUI/QoolColor/CMYKPanel.qml
@@ -1,6 +1,3 @@
-
-pragma ComponentBehavior: Bound
-
 import QtQuick
 import QtQuick.Layouts
 import Qool